feat(Movies): update Vote and Popularity core domain VO

+++ app/src/main/java/org/discovr/mobile/core/domain/Popularity.kt
@@ -1,3 +1,3 @@
 package org.discovr.mobile.core.domain
 
-data class Popularity(val value: Float)
+data class Popularity(val value: Double)
